+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and commits should be formatted using [Conventional Commits](https://www.convent
 
 ### Added
 
-- Config: Migrate (BC for now) from env-only to YAML + env configuration
+- Config: Migrate (BC for now) from env-only to YAML + env configuration ([e7f5cb5](https://github.com/moodlehq/wiki-rag/commit/e7f5cb52b32f90415a53a82c520585138883e6c2))
 
 ### Changed
 
 - GitHub: Update various GH actions to actual versions by @stronk7 ([5851e6a](https://github.com/moodlehq/wiki-rag/commit/5851e6ab0a507d682bb4d9124dd5dd0a08b6d3cd))
+- Config: Re-organise various config sections and model fallbacks
 
 ### Fixed